+++ service/src/main/kotlin/com/example/videostore/movie/MovieController.kt
@@ -2,6 +2,7 @@ package com.example.videostore.movie
 
 import com.example.videostore.movie.dto.MoviePageDTO
 import org.springframework.web.bind.annotation.*
+import java.math.BigDecimal
 
 @CrossOrigin
 @RequestMapping("/movies")
@@ -17,6 +18,6 @@ class MovieController(
     ): MoviePageDTO = movieService.getAllMoviesWith(offset, limit, title)
 
     @PostMapping
-    fun addMovie(@RequestParam(name = "title") title: String, @RequestParam(name = "price") price: Double) =
+    fun addMovie(@RequestParam(name = "title") title: String, @RequestParam(name = "price") price: BigDecimal) =
         movieService.addMovieByTitle(title, price)
 }
